==> CMI7420-2021_Individual_Project/11_SHAP_Analysis_and_Explain-ability.ipynb <==
#import libraries for manipulating and plotting data
import pandas as pd
from scipy import stats

import plotly.graph_objects as go
import plotly.io as pio

pio.templates.default="simple_white"

from progtools.visuals import shap_interaction_plot, time_plot

# 9 Parameter Model

#load the SHAP values in from .csv
df_s_9 = pd.read_csv("SHAP_Values_9_Model.csv",index_col=0)

#import the expected or base value for the 9-parameter model
expected_value_9 = 33.798397

### Mean Absolute Impact on model output magnitude

#create labels to calculate average shap values for physical sensors
labels = ["T50_SHAP","T48_SHAP","P40_SHAP","T30_SHAP","Ps30_SHAP","P50_SHAP","Wf_SHAP","T24_SHAP","P24_SHAP"]

#create an empty list to store average shap values
values = []

#loop through all labels
for i in labels:
    #calculate average shap value for given label
    val = 1/len(df_s_9[i])*df_s_9[i].sum()
    
    #apppend absolute average shap value to list
    values.append(abs(val))

#create a horizontal bar chart of labels and mean absolute SHAP values
fig = go.Figure([go.Bar(y=labels,x=values,orientation="h")])

#update y axes label
fig.update_yaxes(autorange="reversed",title="Sensor")

#update x axes label
fig.update_xaxes(title="mean(|SHAP value|) [average impact on model output magnitude]")
fig.show()

#create labels to calculate average shap values for scenario descriptors
labels = ["TRA_SHAP","alt_SHAP","Mach_SHAP","T2_SHAP"]

#create an empty list to store average shap values
values = []

#loop through all labels
for i in labels:
    #calculate average shap value for given label
    val = 1/len(df_s_9[i])*df_s_9[i].sum()
    
    #apppend absolute average shap value to list
    values.append(abs(val))

#create a horizontal bar chart of labels and mean absolute SHAP values
fig = go.Figure([go.Bar(y=labels,x=values,orientation="h")])

#update y axes label
fig.update_yaxes(autorange="reversed",title="Flight Descriptor")

#update x axes label
fig.update_xaxes(title="mean(|SHAP value|) [average impact on model output magnitude]")
fig.show()

### SHAP Interaction plots

#SHAP interaction plot of T50, T50_SHAP with TRA for unit 1 and cycle 7.
shap_interaction_plot(df_s_9[(df_s_9["unit"]==1)&(df_s_9["cycle"]==7)],
                     "T50",
                     "T50_SHAP",
                     "TRA")

#SHAP interaction plot of T24, T24_SHAP with TRA for unit 1 and cycle 7.
shap_interaction_plot(df_s_9[(df_s_9["unit"]==1)&(df_s_9["cycle"]==7)],
                     "T24",
                     "T24_SHAP",
                     "T48_SHAP")

#time plot of T50 SHAP values at altitidue for unit 1, cycle 7
time_plot(df_s_9[(df_s_9["unit"]==1)&(df_s_9["cycle"]==7)], "alt", "T50_SHAP")

#time plot of T50 SHAP values at altitidue for unit 1, cycle 7
time_plot(df_s_9[(df_s_9["unit"]==1)&(df_s_9["cycle"]==7)], "alt", "T24_SHAP")

**Correlation between T48 and T50 for unit 1, cycle 7:**

#complete Spearman R correlation between T48 and T50 for unit 1, cycle 7
rho, p = stats.spearmanr(df_s_9[(df_s_9["unit"]==1)&(df_s_9["cycle"]==7)]["T48"],
                         df_s_9[(df_s_9["unit"]==1)&(df_s_9["cycle"]==7)]["T50"])

print(f"Test Stat: {rho:.3f} / P-Value: {p:.3f}")

**Correlation between T48 and T50 for unit1, cycle 79:**

#complete Spearman R correlation between T48 and T50 for unit 1, cycle 80
rho, p = stats.spearmanr(df_s_9[(df_s_9["unit"]==1)&(df_s_9["cycle"]==79)]["T48"],
                         df_s_9[(df_s_9["unit"]==1)&(df_s_9["cycle"]==79)]["T50"])

print(f"Test Stat: {rho:.3f} / P-Value: {p:.3f}")

**Correlation between T24 and T48 for unit 1, cycle 7:**

#complete Spearman R correlation between T24 and T50 for unit 1, cycle 7
rho, p = stats.spearmanr(df_s_9[(df_s_9["unit"]==1)&(df_s_9["cycle"]==7)]["T24"],
                         df_s_9[(df_s_9["unit"]==1)&(df_s_9["cycle"]==7)]["T50"])

print(f"Test Stat: {rho:.3f} / P-Value: {p:.3f}")

**Correlation between T24 and T48 for unit 1, cycle 79:**

#complete Spearman R correlation between T24 and T50 for unit 1, cycle 80
rho, p = stats.spearmanr(df_s_9[(df_s_9["unit"]==1)&(df_s_9["cycle"]==79)]["T24"],
                         df_s_9[(df_s_9["unit"]==1)&(df_s_9["cycle"]==79)]["T50"])

print(f"Test Stat: {rho:.3f} / P-Value: {p:.3f}")

**Scatter plot of T50 versus T48, with colours of T50 SHAP values**
<br>The two sensors are more independent, hence provide different information to the algorithm

#interaction plot of T48 and T50 with T50 SHAP values for unit 1, cycle 7
shap_interaction_plot(df_s_9[(df_s_9["unit"]==1)&(df_s_9["cycle"]==7)],
                     "T48",
                     "T50",
                     "T50_SHAP")

**Scatter plot of T50 versus T24, with colours of T50 SHAP values**
<br>It can be seen how closely the two sensors correlate; hence do not provide different information to the algorithm

#interaction plot of T24 and T50 with T50 SHAP values for unit 1, cycle 7
shap_interaction_plot(df_s_9[(df_s_9["unit"]==1)&(df_s_9["cycle"]==7)],
                     "T24",
                     "T50",
                     "T50_SHAP")

**Correlation between T50 and Wf for unit 1, cycle 7:**

#complete Spearman R correlation between Wf and T50 for unit 1, cycle 7
rho, p = stats.spearmanr(df_s_9[(df_s_9["unit"]==1)&(df_s_9["cycle"]==7)]["Wf"],
                         df_s_9[(df_s_9["unit"]==1)&(df_s_9["cycle"]==7)]["T50"])

print(f"Test Stat: {rho:.3f} / P-Value: {p:.3f}")

**Correlation between T50 and Wf for unit 1, cycle 79:**

#complete Spearman R correlation between Wf and T50 for unit 1, cycle 80
rho, p = stats.spearmanr(df_s_9[(df_s_9["unit"]==1)&(df_s_9["cycle"]==79)]["Wf"],
                         df_s_9[(df_s_9["unit"]==1)&(df_s_9["cycle"]==79)]["T50"])

print(f"Test Stat: {rho:.3f} / P-Value: {p:.3f}")

**Scatter plot of T50 versus Wf, with colours of T50 SHAP values**
<br> While there is some non-linearity, providing some fresh information to the algo - they're still very tightly correlated:

#interaction plot of Wf and T50 with T50 SHAP values for unit 1, cycle 7
shap_interaction_plot(df_s_9[(df_s_9["unit"]==1)&(df_s_9["cycle"]==7)],
                     "Wf",
                     "T50",
                     "T50_SHAP")

**Correlation between T50 and P40 for unit 1, cycle 7**

#complete Spearman R correlation between P40 and T50 for unit 1, cycle 7
rho, p = stats.spearmanr(df_s_9[(df_s_9["unit"]==1)&(df_s_9["cycle"]==7)]["P40"],
                         df_s_9[(df_s_9["unit"]==1)&(df_s_9["cycle"]==7)]["T50"])

print(f"Test Stat: {rho:.3f} / P-Value: {p:.3f}")

**Correlation between T50 and P40 for unit 1, cycle 79**

#complete Spearman R correlation between P40 and T50 for unit 1, cycle 79
rho, p = stats.spearmanr(df_s_9[(df_s_9["unit"]==1)&(df_s_9["cycle"]==79)]["P40"],
                         df_s_9[(df_s_9["unit"]==1)&(df_s_9["cycle"]==79)]["T50"])

print(f"Test Stat: {rho:.3f} / P-Value: {p:.3f}")

#interaction plot of P40 and T50 with T50 SHAP values for unit 1, cycle 7
shap_interaction_plot(df_s_9[(df_s_9["unit"]==1)&(df_s_9["cycle"]==7)],
                     "P40",
                     "T50",
                     "T50_SHAP")

#select a unit for analysis

unit = 1
cycle = 79

#create a list of labels for waterfall plot
labels = ["base_value",
          "T50_SHAP","TRA_SHAP","T48_SHAP","P40_SHAP","T30_SHAP","P50_SHAP","T2_SHAP","Ps30_SHAP","alt_SHAP","Mach_SHAP","P24_SHAP","Wf_SHAP","T24_SHAP",
          "Prediction"]

#import base / expected value
values = [expected_value_9]

#cycle through labels, ignoring base_value and prediction
for i in labels:
    if i == "base_value":
        pass
    elif i == "Prediction":
        pass
    else:
        #sum up to 2dp the shap values for that given unit and cycle
        val = round(df_s_9[(df_s_9["unit"]==unit)&(df_s_9["cycle"]==cycle)][i].sum(),2)
        #append the cacluated SHAP value
        values.append(val)

#append a sum of all values in list as the prediction
values.append(round(sum(values),4))

#create a waterfall plot
fig = go.Figure(go.Waterfall(
    #set the orientation to vertical
    orientation="v",
    #set if each value is relative or a total value
    measure = ["relative","relative","relative","relative","relative","relative","relative"
              ,"relative","relative","relative","relative","relative","relative","relative",
              "total"],
    #set the sensor labels as the x axes
    x = labels,
    #set the sensor SHAP values as the y axes
    y = values,
    #include the values as text outside the bars
    text = values,
    textposition = "outside",
    connector = {"line":{"color":"rgb(63,63,63)"}},
))

#update the figure with a suitable title
fig.update_layout(title=f"Waterfall Plot of SHAP Values for Unit {unit}, Cycle {cycle}")
#set the yaxes with a suitable label
fig.update_yaxes(title="RUL Prediction")
fig.show()

# 14 Parameter Model

#load the SHAP values in from .csv
df_s_14 = pd.read_csv("SHAP_Values_14_Model.csv",index_col=0)

#immport expected / base value for prediction
expected_value_14 = 34.000397

#create labels to calculate average shap values for physical sensors
labels = ["T50_SHAP","T48_SHAP","T30_SHAP","Ps30_SHAP","P40_SHAP","P50_SHAP","P2_SHAP",
          "Nf_SHAP","Wf_SHAP", "T24_SHAP","P21_SHAP","Nc_SHAP","P24_SHAP","P15_SHAP"]

#create an empty list to store average shap values
values = []

#loop through all labels
for i in labels:
    #calculate average shap value for given label
    val = 1/len(df_s_14[i])*df_s_14[i].sum()
    
    #apppend absolute average shap value to list
    values.append(abs(val))

#create a horizontal bar chart of labels and mean absolute SHAP values
fig = go.Figure([go.Bar(y=labels,x=values,orientation="h")])

#update y axes label
fig.update_yaxes(autorange="reversed",title="Sensor")

#update x axes label
fig.update_xaxes(title="mean(|SHAP value|) [average impact on model output magnitude]")
fig.show()

#create labels to calculate average shap values for scenario descriptors
labels = ["T2_SHAP","alt_SHAP","TRA_SHAP","Mach_SHAP",]

#create an empty list to store average shap values
values = []

#loop through all labels
for i in labels:
    #calculate average shap value for given label
    val = 1/len(df_s_14[i])*df_s_14[i].sum()
    
    #apppend absolute average shap value to list
    values.append(abs(val))

#create a horizontal bar chart of labels and mean absolute SHAP values
fig = go.Figure([go.Bar(y=labels,x=values,orientation="h")])

#update y axes label
fig.update_yaxes(autorange="reversed",title="Flight Descriptor")

#update x axes label
fig.update_xaxes(title="mean(|SHAP value|) [average impact on model output magnitude]")
fig.show()

### SHAP Interaction plots

#interaction plot of Nf and T50 with T50 SHAP values for unit 1, cycle 5
shap_interaction_plot(df_s_14[(df_s_14["unit"]==1)&(df_s_14["cycle"]==5)],
                     "Nf",
                     "T50",
                     "T50_SHAP")

#interaction plot of P2 and T50 with T50 SHAP values for unit 1, cycle 5
shap_interaction_plot(df_s_14[(df_s_14["unit"]==1)&(df_s_14["cycle"]==5)],
                     "P2",
                     "T50",
                     "P50_SHAP")

#complete Spearman R correlation between Nf and T50 for unit 1, cycle 5
rho, p = stats.spearmanr(df_s_14[(df_s_14["unit"]==1)&(df_s_14["cycle"]==5)]["Nf"],
                         df_s_14[(df_s_14["unit"]==1)&(df_s_14["cycle"]==5)]["T50"])

print(f"Test Stat: {rho:.3f} / P-Value: {p:.3f}")

#complete Spearman R correlation between Nf and T50 for unit 1, cycle 79
rho, p = stats.spearmanr(df_s_14[(df_s_14["unit"]==1)&(df_s_14["cycle"]==79)]["Nf"],
                         df_s_14[(df_s_14["unit"]==1)&(df_s_14["cycle"]==79)]["T50"])

print(f"Test Stat: {rho:.3f} / P-Value: {p:.3f}")

#complete Spearman R correlation between P2 and T50 for unit 1, cycle 5
rho, p = stats.spearmanr(df_s_14[(df_s_14["unit"]==1)&(df_s_14["cycle"]==5)]["P2"],
                         df_s_14[(df_s_14["unit"]==1)&(df_s_14["cycle"]==5)]["T50"])

print(f"Test Stat: {rho:.3f} / P-Value: {p:.3f}")

#complete Spearman R correlation between P2 and T50 for unit 1, cycle 79
rho, p = stats.spearmanr(df_s_14[(df_s_14["unit"]==1)&(df_s_14["cycle"]==79)]["P2"],
                         df_s_14[(df_s_14["unit"]==1)&(df_s_14["cycle"]==79)]["T50"])

print(f"Test Stat: {rho:.3f} / P-Value: {p:.3f}")

### Time plots

#time plot of T50 SHAP values at altitidue for unit 1, cycle 5
time_plot(df_s_14[(df_s_14["unit"]==1)&(df_s_14["cycle"]==5)], "alt", "Nf_SHAP")

#time plot of T50 SHAP values at altitidue for unit 1, cycle 79
time_plot(df_s_14[(df_s_14["unit"]==1)&(df_s_14["cycle"]==79)], "alt", "Nf_SHAP")
